Fix: texture structure separation
