+++ src/base/vgeu_texture.hpp
@@ -18,29 +18,14 @@ https://github.com/SaschaWillems/Vulkan/blob/master/base/VulkanTexture.h
 #define GLM_FORCE_DEPTH_ZERO_TO_ONE
 #include <Vulkan-Hpp/vulkan/vulkan.hpp>
 #include <Vulkan-Hpp/vulkan/vulkan_raii.hpp>
-#define TINYGLTF_IMPLEMENTATION
-#include <tiny_gltf.h>
-#define STB_IMAGE_IMPLEMENTATION
-#define STB_IMAGE_WRITE_IMPLEMENTATION
-#include <stb_image.h>
 
 // std
 #include <memory>
 
 namespace vgeu {
 
-bool isKtx(const tinygltf::Image& gltfImage);
-bool loadImageDataFunc(tinygltf::Image* gltfImage, const int imageIndex,
-                       std::string* error, std::string* warning, int req_width,
-                       int req_height, const unsigned char* bytes, int size,
-                       void* userData);
-bool loadImageDataFuncEmpty(tinygltf::Image* image, const int imageIndex,
-                            std::string* error, std::string* warning,
-                            int req_width, int req_height,
-                            const unsigned char* bytes, int size,
-                            void* userData);
-
-struct Texture {
+class Texture {
+ public:
   std::unique_ptr<vgeu::VgeuImage> vgeuImage;
   vk::ImageLayout imageLayout{};
   uint32_t width = 0;
@@ -50,20 +35,12 @@ struct Texture {
   vk::DescriptorImageInfo descriptorInfo{};
   vk::raii::Sampler sampler = nullptr;
 
-  // fromglTFImage
-  Texture(const tinygltf::Image& gltfimage, const vk::raii::Device& device,
-          VmaAllocator allocator, const vk::raii::Queue& transferQueue,
-          const vk::raii::CommandPool& commandPool);
+  Texture() = default;
   // empty texture
   Texture(const vk::raii::Device& device, VmaAllocator allocator,
           const vk::raii::Queue& transferQueue,
           const vk::raii::CommandPool& commandPool);
 
-  // TODO: move into gltf using inheritance
-  void fromglTFImage(const tinygltf::Image& gltfimage,
-                     const vk::raii::Device& device, VmaAllocator allocator,
-                     const vk::raii::Queue& transferQueue,
-                     const vk::raii::CommandPool& commandPool);
   void generateMipmaps(const vk::raii::CommandBuffer& cmdBuffer);
   void createEmptyTexture(const vk::raii::Device& device,
                           VmaAllocator allocator,
